+++ pandas_vet/SeriesVet.py
@@ -458,7 +458,7 @@ class SeriesVet:
 
     def print_time_elapsed(
         self,
-        start_time: int,
+        start_time: float,
         lead_in: Union[str, None] = "Time elapsed",
         units: str = "auto",
     ) -> pd.Series:
